pyproject.toml
[tool.poetry]
name = "vivli"
version = "0.1.0"
description = "avito analog for sakartvelo"
authors = ["FourZd <shogo.ryaskov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["websockets"], version = "^0.31.1"}
dependency-injector = "^4.42.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pre-commit = "^3.8.0"
hatchling = "^1.25.0"
alembic = "^1.13.3"
pydantic-settings = "^2.5.2"
psycopg2 = "^2.9.9"
pyjwt = "^2.9.0"
bcrypt = "^4.2.0"
aiosmtplib = "^3.0.2"
redis = "^5.1.0"
asyncpg = "^0.29.0"
isort = "^5.13.2"
email-validator = "^2.2.0"
celery = {extras = ["asyncio"], version = "^5.4.0"}
geopy = "^2.4.1"
python-multipart = "^0.0.12"
asgiref = "^3.8.1"
nest-asyncio = "^1.6.0"
python-socketio = {extras = ["redis"], version = "^5.11.4"}
babel = "^2.16.0"
firebase-admin = "^6.5.0"
authlib = "^1.3.2"
httpx = "^0.27.2"
itsdangerous = "^2.2.0"
phonenumbers = "^8.13.52"
yookassa = "^3.5.0"
pillow = "^11.1.0"
pillow-heif = "^0.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"